backend/requirements.txt
fastapi
uvicorn
python-dotenv
pymongo
pydantic
pillow
